Update utils.py
Casting error for imshow xtic, ytic

+++ utils.py
@@ -22,14 +22,14 @@ def imshow(im, x_vec=None, y_vec=None, dynamic_range=70):
     plt.imshow(img, cmap='gray')
     
     try:
-        x_tic = np.linspace(x_vec[0], x_vec[-1], 11)
+        x_tic = np.linspace(x_vec[0], x_vec[-1], 11, dtype=np.int32)
         x_loc = np.linspace(0, img.shape[0], 11, dtype=np.int32)
         plt.xticks(x_loc, x_tic)
     except TypeError:
         pass
     
     try:
-        y_tic = np.linspace(y_vec[0], y_vec[-1], 11)
+        y_tic = np.linspace(y_vec[0], y_vec[-1], 11, dtype=np.int32)
         y_loc = np.linspace(0, img.shape[1], 11, dtype=np.int32)
         plt.yticks(y_loc, y_tic)
     except TypeError:
